requirements.txt
annotated-types==0.5.0
anyio==4.0.0
assemblyai==0.17.0
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
websockets==11.0.3
